+++ hipparchus-core/src/main/java/org/hipparchus/fraction/Fraction.java
@@ -372,7 +372,7 @@ public class Fraction
      */
     @Override
     public double doubleValue() {
-        return (double)numerator / (double)denominator;
+        return ((double) numerator) / denominator;
     }
 
     /**
